Working Tags
This commit adds a small Ruby “Tag Generator” plugin. It’s not fancy,
but it gets the job done. Also, I made sure that the small “tag
selector” in the header works. Finally, to get it all to work, I had to
change some of the tags in the posts.

Closes #98